# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/member": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查詢 Member 清單，支援關鍵字與排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "取得 Member 列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼（預設 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（預設 10）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "關鍵字搜尋",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序欄位與方式，如 created_at desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/member/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 取得 Member 詳細資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "取得單一 Member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳 User",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 修改 Member 權限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "修改 Member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的 Member 資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MenberUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳更新後的 User",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/api/todo/details": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "建立一個新的 todoListDetails 項目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoListDetails"
                ],
                "summary": "新增 todoListDetails",
                "responses": {
                    "200": {
                        "description": "建立成功回傳的 TodoListDetails 資料",
                        "schema": {
                            "$ref": "#/definitions/models.TodoListDetails"
                        }
                    }
                }
            }
        },
        "/api/todo/details/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 修改 TodoListDetails 名稱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoListDetails"
                ],
                "summary": "修改 TodoListDetails",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoListDetails ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的 TodoListDetails 資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TodoTypeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳更新後的 TodoType",
                        "schema": {
                            "$ref": "#/definitions/models.TodoListDetails"
                        }
                    }
                }
            }
        },
        "/api/todo/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查詢 TodoList 清單，支援關鍵字與排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList"
                ],
                "summary": "取得 TodoList 列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼（預設 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（預設 10）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "關鍵字搜尋",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序欄位與方式，如 created_at desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "建議一個新的 todoList 項目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList"
                ],
                "summary": "新增 TodoList",
                "parameters": [
                    {
                        "description": "建立 TodoList 所需資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TodoListCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "建立成功回傳的 TodoList 資料",
                        "schema": {
                            "$ref": "#/definitions/models.TodoList"
                        }
                    }
                }
            }
        },
        "/api/todo/list/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 取得 TodoList 詳細資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList"
                ],
                "summary": "取得單一 TodoList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳 TodoList",
                        "schema": {
                            "$ref": "#/definitions/models.TodoList"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 修改 TodoList 名稱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList"
                ],
                "summary": "修改 TodoList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的 TodoList 資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TodeListUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳更新後的 TodoList",
                        "schema": {
                            "$ref": "#/definitions/models.TodoList"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 刪除指定的 TodoList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoList"
                ],
                "summary": "刪除 TodoList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳被刪除的 TodoList",
                        "schema": {
                            "$ref": "#/definitions/models.TodoList"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.MenberUpdate": {
            "type": "object",
            "required": [
                "role_id"
            ],
            "properties": {
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "dto.TodeListUpdateRequest": {
            "type": "object",
            "required": [
                "name",
                "type_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "type_id": {
                    "type": "integer"
                }
            }
        },
        "dto.TodoListCreateRequest": {
            "type": "object",
            "required": [
                "name",
                "type_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "type_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TodoTypeUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.TodoList": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TodoListDetails"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TodoTypes"
                },
                "type_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.TodoListDetails": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "to_do_list_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.TodoTypes": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "account"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "todo_list_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TodoListDetails"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        }
    }
}